Project done v.2

+++ app/Http/Controllers/LogoController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use App\Models\Logo;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Storage;
+use Image;
 
 class LogoController extends Controller
 {
@@ -76,11 +77,15 @@ class LogoController extends Controller
      */
     public function update(Request $request, Logo $logo)
     {
-        $updateEntry = $logo;
-        Storage::disk('public')->delete('img/'.$logo->src);
-        $request->file('src')->storePublicly('img/', 'public');
-        $updateEntry->src = $request->file('src')->hashName();
-        $updateEntry->save();
+        if ($request->file('src') != NULL) {
+            $updateEntry = $logo;
+            Storage::disk('public')->delete('img/'.$logo->src);
+            $request->file('src')->storePublicly('img/', 'public');
+            $updateEntry->src = $request->file('src')->hashName();
+            $updateEntry->save();
+            $img = Image::make('img/'.$updateEntry->src)->resize(100,80);
+            $img->save('img/small-'.$updateEntry->src);
+        }
         return redirect()->back();
     }
